Rename `newtype.rs` to `paths.rs`
